Updated the content to fit the newest information
Much was changed, such as the Mention System is now in V.3 and has more features, which has now been included. Discord Buttons have been announced and introduced for Tsumiki, which also is now included. Information has been partially or fully updated accordingly with what's new or changed with Tsumiki since the last Github README.md push.
Compiled the text more and made it more compact with the use of \ at the end of sentences I want to retain the paragraph feeling. And so much more was also changed and adjusted. See comparison for the entire list of changes made.