Changed Bot name & Updated to fit new theme
The newest release of a complete revamp of the previously bot known as "Tsumiki" is a new and improved system that will allow me to completely push a single button and change the bot's name, and the new system will follow suit. The bot is now completely autonomous when it comes to drastic and major changes. Over 30,000+ lines of code have been altered in some shape or form and all internal mentions are now completely gone.